add experimental "npm run previewOnly"
