+++ public/src/machines/fetchMachine.js
@@ -1,0 +1,38 @@
+
+import {assign, Machine} from 'xstate'
+
+export const fetchMaschine = Machine({
+    id: 'fetch',
+    initial: 'idle',
+    context: {
+      retries: 0
+    },
+    states: {
+      idle: {
+        on: {
+          FETCH: 'loading'
+        }
+      },
+      loading: {
+        on: {
+          RESOLVE: 'success',
+          REJECT: 'failure'
+        }
+      },
+      success: {
+        type: 'final'
+      },
+      failure: {
+        on: {
+          RETRY: {
+            target: 'loading',
+            actions: assign({
+              retries: context => context.retries + 1
+            })
+          }
+        }
+      }
+    }
+  })
+  
+
